DevCommit #26.11.20.1
GameData: added information about the protected cells, on which construction is prohibited.
AttackEnemy, EnemyController: constructors changed. Now it take object called _stat and assign all properties of this object to the parent.
HealthBar: is moved to a separate file.

Now enemies vary in type and gems vary in target type.
Added a config file for enemies.
Replacing some sprites.