DevCommit #25.12.20.1
The new class State, contains the properties of the object (health, speed, damage) and their options such as the maximum and minimum values (for numbers). The class automatically creates setters and getters for the added properties. Setters make sure that numerical values do not go beyond the specified range (max, min).

MoveController, AttackEnemy, EnemyController, HealthBar are configured to work with the State class.