DevCommit #05.01.21.1
New class EventController is a local event system for object. Allows to add and subscribe to events that occur in the components of an object.
Info list display effects and skills of objects.

Restructured data file - gems.json.
The system of effects has been improved.
Added new skills and effects